requirements.txt
# setuptools==65.5.0 # arquitetura
duckdb==1.4.3
magic_duckdb==0.2.4
mysql-connector==2.2.9
nbformat==5.10.4
numpy==2.4.1
openpyxl==3.1.5
pandas==2.3.3
pyarrow==23.0.0
python-decouple==3.8
sqlalchemy==2.0.45